app/requirements.txt
fastapi
uvicorn
python-dotenv
PyPDF2
google-generativeai
pollinations
